rename mongoDBHandler package
